Refactored login hint process
