+++ app/authenticators/gdrive.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import Base from 'simple-auth/authenticators/base';
 import Auth from 'ember-gdrive/lib/auth';
-import Cache from 'ember-gdrive/lib/local-cache';
+import { fetchLoginHint } from 'ember-drive/lib/login-hint';
 import config from 'ember-gdrive/lib/config';
 
 var Authenticator = Base.extend({
@@ -13,7 +13,7 @@ var Authenticator = Base.extend({
   restore: function (properties) {
     var authenticator = this;
     return authenticator.get('auth').authorizeImmediate({
-      login_hint: this.inferUserId(),
+      login_hint: fetchLoginHint(),
       client_id: config.get('GOOGLE_CLIENT_ID')
     }).then(function () {
       return authenticator.get('auth').fetchCurrentUser();
@@ -30,27 +30,6 @@ var Authenticator = Base.extend({
 
   invalidate: function () {
     return this.get('auth').close();
-  },
-
-  extractQueryParams: function () {
-    var params = {};
-    location.search.substr(1).split('&').forEach(function (item) {
-      params[item.split('=')[0]] = decodeURIComponent(item.split('=')[1]);
-    });
-    return params.state ? JSON.parse(params.state) : {};
-  },
-
-  getDocumentIdFromLocation: function () {
-    return location.href.split('/d/')[1].split('/')[0];
-  },
-
-  inferUserId: function() {
-    var userId = this.extractQueryParams().userId;
-    if (!userId) {
-      var cache = new Cache('document_login_hint');
-      userId = cache.get(this.getDocumentIdFromLocation());
-    }
-    return userId;
   }
 });
