--- design_notes.txt
@@ -1,5 +1,0 @@
-create file
-?state=%7B%22folderId%22:%220AMftiYAzT3YQUk9PVA%22,%22action%22:%22create%22,%22userId%22:%117742084218787125292%22%7D
-
-open file
-?state=%7B%22ids%22:%5B%220B8ftiYAzT3YQcDc5V3RScjl3OU0%22%5D,%22action%22:%22open%22,%22userId%22:%22117742084218787125292%22%7D
